requirements.txt
langchain
langgraph
langchain-groq
langchain-google-genai
python-dotenv
tavily-python
playwright
pandas
plotly
fastapi
uvicorn